refactor: add manual implementation to example app
- Removed unused imports and code from App.tsx
- Introduced tab navigation to switch between Hooks and Manual screens
- Created separate components for AppHashSection, InstructionsSection, and StatusMessages
- Moved SMS listener logic to Manual screen with direct module usage
- Added styles for tab navigation and improved code organization
- Created types for StatusMessage to enhance type safety
