+++ example/types.ts
@@ -1,0 +1,5 @@
+export type StatusMessage = {
+  type: 'success' | 'error' | 'info' | 'warning'
+  message: string
+  timestamp: number
+}
